[OZON-3], код ревью, добавил валидацию для value object, добавил тесты для value object

+++ src/OzonEdu.MerchandiseService.Domain/AggregationModels/IssuedMerchAggregate/IIssuedMerchRepository.cs
@@ -17,7 +17,7 @@ namespace OzonEdu.MerchandiseService.Domain.AggregationModels.IssuedMerchAggrega
 		/// <param name="ct">Токен для отмены операции <see cref="CancellationToken"/></param>
 		/// <returns>Список сущностей <see cref="IssuedMerch"/>.</returns>
 		Task<IReadOnlyCollection<IssuedMerch>> GetIssuedMerches(
-			int employeeId,
+			long employeeId,
 			CancellationToken ct = default);
 	}
 }
